MGDOBR-187: move exceptions to infra-api
