+++ actions/src/test/java/com/redhat/service/bridge/actions/ActionProviderFactoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 
 import com.redhat.service.bridge.actions.kafkatopic.KafkaTopicAction;
 import com.redhat.service.bridge.actions.webhook.WebhookAction;
+import com.redhat.service.bridge.infra.api.exceptions.ActionProviderException;
 
 import io.quarkus.test.junit.QuarkusTest;
