+++ infra/infra-api/src/main/java/com/redhat/service/bridge/infra/api/exceptions/ActionProviderException.java
@@ -1,4 +1,4 @@
-package com.redhat.service.bridge.actions;
+package com.redhat.service.bridge.infra.api.exceptions;
 
 public class ActionProviderException extends RuntimeException {
