+++ infra/infra-api/src/main/java/com/redhat/service/bridge/infra/api/exceptions/AlreadyExistingItemException.java
@@ -1,16 +1,16 @@
-package com.redhat.service.bridge.manager.exceptions;
+package com.redhat.service.bridge.infra.api.exceptions;
 
 import javax.ws.rs.core.Response;
 
-public class BridgeLifecycleException extends EventBridgeManagerException {
+public class AlreadyExistingItemException extends EventBridgeException {
 
     private static final long serialVersionUID = 1L;
 
-    public BridgeLifecycleException(String message) {
+    public AlreadyExistingItemException(String message) {
         super(message);
     }
 
-    public BridgeLifecycleException(String message, Throwable cause) {
+    public AlreadyExistingItemException(String message, Throwable cause) {
         super(message, cause);
     }
