+++ infra/infra-api/src/main/java/com/redhat/service/bridge/infra/api/exceptions/BridgeLifecycleException.java
@@ -1,16 +1,16 @@
-package com.redhat.service.bridge.manager.exceptions;
+package com.redhat.service.bridge.infra.api.exceptions;
 
 import javax.ws.rs.core.Response;
 
-public class AlreadyExistingItemException extends EventBridgeManagerException {
+public class BridgeLifecycleException extends EventBridgeException {
 
     private static final long serialVersionUID = 1L;
 
-    public AlreadyExistingItemException(String message) {
+    public BridgeLifecycleException(String message) {
         super(message);
     }
 
-    public AlreadyExistingItemException(String message, Throwable cause) {
+    public BridgeLifecycleException(String message, Throwable cause) {
         super(message, cause);
     }
