+++ ingress/src/main/java/com/redhat/service/bridge/ingress/api/IngressAPI.java
@@ -18,7 +18,7 @@ import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.redhat.service.bridge.infra.api.utils.CloudEventUtils;
-import com.redhat.service.bridge.ingress.api.exceptions.BadRequestException;
+import com.redhat.service.bridge.infra.api.exceptions.BadRequestException;
 import com.redhat.service.bridge.ingress.producer.KafkaEventPublisher;
 
 import io.cloudevents.CloudEvent;
