+++ manager/src/test/java/com/redhat/service/bridge/manager/api/user/validators/actions/ActionParamValidatorContainerTest.java
@@ -12,9 +12,9 @@ import org.mockito.ArgumentCaptor;
 
 import com.redhat.service.bridge.actions.ActionParameterValidator;
 import com.redhat.service.bridge.actions.ActionProvider;
-import com.redhat.service.bridge.actions.ActionProviderException;
 import com.redhat.service.bridge.actions.ActionProviderFactory;
 import com.redhat.service.bridge.actions.ValidationResult;
+import com.redhat.service.bridge.infra.api.exceptions.ActionProviderException;
 import com.redhat.service.bridge.infra.api.models.actions.BaseAction;
 import com.redhat.service.bridge.manager.api.models.requests.ProcessorRequest;
